Plots Kriged surfaces of functional characteristics

plot.kriged_surfaces() produces spatial landscape plots of kriged surfaces produced by krige_surf.

Usage

# S3 method for class 'kriged_surfaces'
plot(x, alpha = 0.5, pt.col = "black",
     interpolate = TRUE, contour = TRUE, ...)

Arguments

x

a kriged_surfaces object; the output of a call to krige_surf.

alpha, pt.col

when a new_data component is present in x, the transparency (alpha) and color (pt.col) of the points plotted for the new samples.

interpolate

logical; whether to smooth the plot by interpolating across pixels in the grid. Passed to ggplot2::geom_raster.

contour

logical; whether to add contour lines to the plot to illustrate changes in the fitness landscape.

...

ignored.

Details

plot.kriged_surfaces() is a wrapper for ggplot2 raster plotting functions. For more precise control of raster plotting see ggplot2::geom_raster.

Value

A ggplot object, which can be further manipulated using ggplot2 functionality.

See also

Examples

# See examples at help("krige_surf")